add cache switch

+++ inference/sdxl_infer.py
@@ -46,9 +46,12 @@ def sdxl_infer(prompt: str,
                width: int = 1024,
                scale: float = 10,
                steps: int = 50,
-               seed: int = 0):
-    cache_dir = get_cache_dir()
-    shutil.rmtree(cache_dir)
+               seed: int = 0,
+               clear_cache: bool = False):
+    if clear_cache:
+        print("Clear download model weights.")
+        cache_dir = get_cache_dir()
+        shutil.rmtree(cache_dir)
     pipe = pipeline(task=Tasks.text_to_image_synthesis,
                     model='AI-ModelScope/stable-diffusion-xl-base-1.0',
                     use_safetensors=True,
